refactor: separate all the reviews
